Fix stale rope delta issue in dflash pipeline
