+++ src/cpp/src/speculative_decoding/continuous_batching/eagle3_strategy.hpp
@@ -87,7 +87,6 @@ protected:
     // the wrapper for executing kv cache update model in eagle3 pipeline
     std::shared_ptr<KVUpdateWrapper> m_kv_update_wrapper;
     ov::Tensor create_draft_input_embeddings(const ov::Tensor& original_input_embeddings);
-    static int64_t compute_rope_delta(const ov::Tensor& position_ids);
     static ov::Tensor trim_first_token_sequence_tensor(const ov::Tensor& tensor, const char* tensor_name);
 };
 }
